Get rid of CommonOperator
